refactor logic - removing the variant check from constants module and change to a server side test
